HSEARCH-3476 Add one SearchPredicate implementation per Lucene predicate
Instead of the previous hack consisting in shoving both the building
behavior and the predicate behavior into the builder in order to have
a single LuceneSearchPredicate implementation that delegates
to the builder.

There are two advantages to this solution:

1. It's much easier to ensure that predicates are not mutated after the
   fact, since we copy the data from the builder.
2. We no longer need to reference the ElasticsearchSearchContext from
   predicates, even indirectly, meaning the context and related objects
   can be garbage collected even if the predicate is cached by the user.
