--- backend/lucene/src/main/java/org/hibernate/search/backend/lucene/search/predicate/impl/AbstractLuceneSpatialWithinBoundingBoxPredicateBuilder.java
@@ -1,30 +1,0 @@
-/*
- * Hibernate Search, full-text search for your domain model
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.search.backend.lucene.search.predicate.impl;
-
-import org.hibernate.search.backend.lucene.search.impl.LuceneSearchContext;
-import org.hibernate.search.backend.lucene.search.impl.LuceneSearchFieldContext;
-import org.hibernate.search.engine.search.predicate.spi.SpatialWithinBoundingBoxPredicateBuilder;
-import org.hibernate.search.engine.spatial.GeoBoundingBox;
-
-
-public abstract class AbstractLuceneSpatialWithinBoundingBoxPredicateBuilder
-		extends AbstractLuceneSingleFieldPredicateBuilder
-		implements SpatialWithinBoundingBoxPredicateBuilder {
-
-	protected GeoBoundingBox boundingBox;
-
-	protected AbstractLuceneSpatialWithinBoundingBoxPredicateBuilder(LuceneSearchContext searchContext,
-			LuceneSearchFieldContext<?> field) {
-		super( searchContext, field );
-	}
-
-	@Override
-	public void boundingBox(GeoBoundingBox boundingBox) {
-		this.boundingBox = boundingBox;
-	}
-}
