+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/search/predicate/impl/LuceneMatchIdPredicate.java
@@ -14,33 +14,24 @@ import org.hibernate.search.backend.lucene.search.impl.LuceneSearchContext;
 import org.hibernate.search.engine.backend.types.converter.runtime.spi.ToDocumentIdentifierValueConvertContext;
 import org.hibernate.search.engine.backend.types.converter.spi.ToDocumentIdentifierValueConverter;
 import org.hibernate.search.engine.search.common.ValueConvert;
+import org.hibernate.search.engine.search.predicate.SearchPredicate;
 import org.hibernate.search.engine.search.predicate.spi.MatchIdPredicateBuilder;
 
 import org.apache.lucene.index.Term;
 import org.apache.lucene.search.BooleanClause.Occur;
 import org.apache.lucene.search.BooleanQuery;
-import org.apache.lucene.search.BooleanQuery.Builder;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.TermQuery;
 
-public class LuceneMatchIdPredicateBuilder extends AbstractLuceneSearchPredicateBuilder
-		implements MatchIdPredicateBuilder {
+public class LuceneMatchIdPredicate extends AbstractLuceneSearchPredicate {
 
-	private final List<String> values = new ArrayList<>();
-	private final LuceneSearchContext searchContext;
+	private final List<String> values;
 
-	LuceneMatchIdPredicateBuilder(LuceneSearchContext searchContext) {
-		super( searchContext );
-		this.searchContext = searchContext;
-	}
-
-	@Override
-	public void value(Object value, ValueConvert valueConvert) {
-		ToDocumentIdentifierValueConverter<?> dslToDocumentIdConverter =
-				searchContext.indexes().idDslConverter( valueConvert );
-		ToDocumentIdentifierValueConvertContext toDocumentIdentifierValueConvertContext =
-				searchContext.toDocumentIdentifierValueConvertContext();
-		values.add( dslToDocumentIdConverter.convertUnknown( value, toDocumentIdentifierValueConvertContext ) );
+	private LuceneMatchIdPredicate(Builder builder) {
+		super( builder );
+		values = builder.values;
+		// Ensure illegal attempts to mutate the predicate will fail
+		builder.values = null;
 	}
 
 	@Override
@@ -49,8 +40,8 @@ public class LuceneMatchIdPredicateBuilder extends AbstractLuceneSearchPredicate
 	}
 
 	@Override
-	protected Query doBuild(PredicateRequestContext context) {
-		Builder builder = new BooleanQuery.Builder();
+	protected Query doToQuery(PredicateRequestContext context) {
+		BooleanQuery.Builder builder = new BooleanQuery.Builder();
 		for ( String value : values ) {
 			builder.add( termQuery( value ), Occur.SHOULD );
 		}
@@ -59,5 +50,27 @@ public class LuceneMatchIdPredicateBuilder extends AbstractLuceneSearchPredicate
 
 	private TermQuery termQuery( String value ) {
 		return new TermQuery( new Term( MetadataFields.idFieldName(), value ) );
+	}
+
+	static class Builder extends AbstractBuilder implements MatchIdPredicateBuilder {
+		private List<String> values = new ArrayList<>();
+
+		Builder(LuceneSearchContext searchContext) {
+			super( searchContext );
+		}
+
+		@Override
+		public void value(Object value, ValueConvert valueConvert) {
+			ToDocumentIdentifierValueConverter<?> dslToDocumentIdConverter =
+					searchContext.indexes().idDslConverter( valueConvert );
+			ToDocumentIdentifierValueConvertContext toDocumentIdentifierValueConvertContext =
+					searchContext.toDocumentIdentifierValueConvertContext();
+			values.add( dslToDocumentIdConverter.convertUnknown( value, toDocumentIdentifierValueConvertContext ) );
+		}
+
+		@Override
+		public SearchPredicate build() {
+			return new LuceneMatchIdPredicate( this );
+		}
 	}
 }
